+++ Quest Tracker/CoreDataObjects/Settings+CoreDataClass.swift
