+++ Quest Tracker/CoreDataObjects/Settings+CoreDataProperties.swift
